% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-content.R
\name{compose_content_request}
\alias{compose_content_request}
\title{Create a content request}
\usage{
compose_content_request(user, repo, path, ...)
}
\arguments{
\item{user}{The user name under which the repo of interest lives.}

\item{repo}{The name of the repository of interest.}

\item{path}{The path to the file for which the content should be retrieved
relative to the root of `repo`.}

\item{...}{Parameters for the request in the form of key / value parirs. See
also [key_value_pair_to_chr()]}
}
\description{
Create a content request
}
