% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespaces-parse.R
\name{expand_multiple_io_per_line}
\alias{expand_multiple_io_per_line}
\title{Expand NAMESPACE to one entry per line}
\usage{
expand_multiple_io_per_line(tabular_namespace)
}
\arguments{
\item{tabular_namespace}{A name space in a tabular format, that is, a tibble
with a column type and one with object.}
}
\description{
Expand a line from a NAMESPACE listing multiple objects into a multiple lines
containing one object each line.
}
\examples{
tabular <- tibble::tibble(
  type = "export",
  object = "base64encode, base64decode, dataURI, checkUTF8"
)
expand_multiple_io_per_line(tabular)
}
