% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-content.R
\name{fetch_gh_content}
\alias{fetch_gh_content}
\title{Fetch the content of a GitHub file}
\usage{
fetch_gh_content(user, repo, path, ...)
}
\arguments{
\item{user}{The user name under which the repo of interest lives.}

\item{repo}{The name of the repository of interest.}

\item{path}{The path to the file for which the content should be retrieved
relative to the root of `repo`.}

\item{...}{Parameters passed to [compose_content_request()].}
}
\description{
Get the content of a file from a GitHub repository.
}
\examples{
fetch_gh_content(
  user = "cran", repo = "styler", path = "NAMESPACE", ref = "1.0.0"
)
}
