% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-tags.R
\name{fetch_gh_tags}
\alias{fetch_gh_tags}
\title{Fetch all tags of a github repo}
\usage{
fetch_gh_tags(user, repo)
}
\description{
Fetch all tags of a github repo
}
\examples{
fetch_gh_tags("cran", "styler")
}
