% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace-collect.R
\name{find_first_release}
\alias{find_first_release}
\title{Which release is the first to export something?}
\usage{
find_first_release(package, object)
}
\arguments{
\item{package}{A package.}

\item{object}{The exported object  in \code{package}. If \code{NA}, simply the last
release is returned.}
}
\description{
Find the first release of \code{package} that exports \code{object}. If no match is
found, \code{NA} is returned.
}
\details{
Queries the \href{https://github.com/metacran}{GitHub CRAN mirror} to find all
releases of a packages and performs binary search on their NAMESPACE to find
the first release in which a certain object was exported.
}
\examples{
find_first_release("tibble", "add_column")
find_first_release("tibble", "add_columns")
}
