% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisect.R
\name{may_terminate}
\alias{may_terminate}
\title{How did it all end?}
\usage{
may_terminate(range, fun)
}
\arguments{
\item{range}{A range of candidates for the bisection}

\item{fun}{A function to be applied to range once range shrank to length one.}
}
\description{
Asssess the result of a binary search. Return the successful condition
if one is found, otherwise return NA. If the search has not yet been
determined, return a character string of length zero. The function is not
written for performance, but just to be explicit.
}
\keyword{internal}
