% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{may_unlist}
\alias{may_unlist}
\title{Wrapper around tibble::deframe()}
\usage{
may_unlist(x, deframe)
}
\arguments{
\item{x}{object to deframe}

\item{deframe}{Whether or not to deframe.}
}
\description{
Wrapper around tibble::deframe()
}
