% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespaces-package.R
\docType{package}
\name{namespaces-package}
\alias{namespaces}
\alias{namespaces-package}
\title{Working with name spaces}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Lorenz Walthert \email{lorenz.walthert@icloud.com}

}
\keyword{internal}
