% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace-collect.R
\name{obj_is_exported_in_ref}
\alias{obj_is_exported_in_ref}
\title{Check whether an oject is exported by a certain package version}
\usage{
obj_is_exported_in_ref(package, ref, obj)
}
\arguments{
\item{package}{The package to check.}

\item{ref}{The version reference to check.}

\item{obj}{The object for which we want to know whether it is exported or
not.}
}
\description{
Check whether an oject is exported by a certain package version
}
