% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespaces-io.R
\name{parse_namespace}
\alias{parse_namespace}
\alias{parse_local_namespace}
\alias{parse_cran_namespace}
\title{Parse a namespace}
\usage{
parse_local_namespace(path = ".")

parse_cran_namespace(package, ...)
}
\arguments{
\item{path}{The path to the source code of a package.}

\item{package}{The name of a package available on CRAN.}
}
\description{
Parses a NAMESPACE file from a package into a tabular format.
}
\section{Functions}{
\itemize{
\item \code{parse_local_namespace}: Parses a local name space.

\item \code{parse_cran_namespace}: Parses a namespace from the GitHub cran mirror.
}}

\examples{
random_dir <- tempdir()
pkg_dir <- file.path(random_dir, "randompkg")
dir.create(pkg_dir)
enc::write_lines_enc(c(
  "# Generated by roxygen2: do not edit by hand",
  "",
  "export(parse_cran_namespace)",
  "export(parse_local_namespace)",
  "import(gh)",
  "import(tibble)",
  "importFrom(magrittr,\\"\%>\%\\")",
  "importFrom(purrr,map_chr)",
  "importFrom(purrr,partial)"
), file.path(pkg_dir, "NAMESPACE"))
parse_local_namespace(pkg_dir)
unlink(random_dir)
parse_cran_namespace("tibble")
parse_cran_namespace("styler", ref = "1.0.0")
}
