% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespaces-package.R
\docType{package}
\name{namespaces-package}
\alias{namespaces}
\alias{namespaces-package}
\title{Working with name spaces}
\description{
Working with name spaces
}
\author{
\strong{Maintainer}: Lorenz Walthert \email{lorenz.walthert@icloud.com}

}
\keyword{internal}
