% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespaces-io.R
\name{write_minimal_dependencies}
\alias{write_minimal_dependencies}
\title{Write out minimal dependencies from a tabular namespace}
\usage{
write_minimal_dependencies(path = ".", namespace = NULL)
}
\arguments{
\item{path}{The path to the source package for which DESCRIPTION should be
adapted.}

\item{namespace}{A name space in a tabular format, i.e. a data frame with
two columns \code{package} and \code{object} to be written to a file. If \code{NULL},
\code{\link[=collect_minimal_dependencies]{collect_minimal_dependencies()}} will be used to obtain the information
necessary.}
}
\description{
Write out minimal dependencies from a tabular namespace
}
